% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{generate_samples}
\alias{generate_samples}
\title{Generate samples from a model}
\usage{
generate_samples(experiment, n, eta = NULL)
}
\arguments{
\item{experiment}{the experiment data returned by the [run_experiment()] function.}

\item{n}{the number of samples to generate}

\item{eta}{the number of iterations in the SGLD sampler.}
}
\description{
Generate samples from a model
}
