% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_samples}
\alias{plot_samples}
\title{Plot samples}
\usage{
plot_samples(samps, nrow = 8)
}
\arguments{
\item{samps}{a 4D tensor of samples (as generated by the [generate_samples()])
function.}

\item{nrow}{number of rows in the generated figure.}
}
\description{
Plot generated samples
}
