% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{run_experiment}
\alias{run_experiment}
\title{Run experiment}
\usage{
run_experiment(config = config::get())
}
\arguments{
\item{config}{a config list with the information below. If empty, will use
the config.yml in the working directory.}
}
\value{
Returns all experiment information. Losses, model, buffer and configuration.
}
\description{
Trains annd validates the JEM model using the specified config.
}
